// types/sanityTypes.ts
export interface ListItem {
  item: string;
  nestedList?: ListItem[];
}

export interface ImageWithAlt {
  imageUrl?: SanityImageSource;
  alt?: string;
}
// ...
export interface LeftRightData {
  leftData?: {
    title?: string;
    lists?: ListItem[];
  };
  rightData?: ImageWithAlt;
}

export interface TravelVisasSection {
  mainTitle?: string;
  mainDesc?: string;
  section1?: LeftRightData;
  section2?: LeftRightData;
}